requirements.txt
streamlit>=1.35,<1.40
sqlmodel>=0.0.16
sqlalchemy>=2.0
pydantic>=2.7
pydantic-settings>=2.4
alembic>=1.13
networkx>=3.3
pyvis>=0.3.2
requests>=2.32
python-dotenv>=1.0
python-docx>=1.1
python-pptx>=0.6.23
pdfplumber>=0.11
pypdf>=4.2
openpyxl>=3.1
pandas>=2.2
uvicorn>=0.30
tenacity>=8.2
orjson>=3.10
rapidfuzz>=3.8
tiktoken>=0.7
